Dev: move rolling average utils to Transforms and start transforms test file
