+++ coreTable/Transforms.ts
@@ -1,11 +1,93 @@
-import {
-    computeRollingAverage,
-    insertMissingValuePlaceholders,
-} from "explorer/covidExplorer/CovidExplorerUtils"
 import { flatten } from "grapher/utils/Util"
 import { ColumnSlug, CoreColumnStore, Time } from "./CoreTableConstants"
 import { CoreColumnDef } from "./CoreColumnDef"
 import { InvalidCell, InvalidCellTypes, isValid } from "./InvalidCells"
+
+// In Grapher we return just the years for which we have values for. This puts MissingValuePlaceholder
+// in the spots where we are missing values (added to make computing rolling windows easier).
+// Takes an array of value/year pairs and expands it so that there is an undefined
+// for each missing value from the first year to the last year, preserving the position of
+// the existing values.
+export function insertMissingValuePlaceholders(
+    values: number[],
+    times: number[]
+) {
+    const startTime = times[0]
+    const endTime = times[times.length - 1]
+    const filledRange = []
+    let time = startTime
+    const timeToValueIndex = new Map()
+    times.forEach((time, index) => {
+        timeToValueIndex.set(time, index)
+    })
+    while (time <= endTime) {
+        filledRange.push(
+            timeToValueIndex.has(time)
+                ? values[timeToValueIndex.get(time)]
+                : InvalidCellTypes.MissingValuePlaceholder
+        )
+        time++
+    }
+    return filledRange
+}
+
+// todo: add the precision param to ensure no floating point effects
+export function computeRollingAverage(
+    numbers: (number | undefined | null | InvalidCell)[],
+    windowSize: number,
+    align: "right" | "center" = "right"
+) {
+    const result: (number | InvalidCell)[] = []
+
+    for (let valueIndex = 0; valueIndex < numbers.length; valueIndex++) {
+        // If a value is undefined in the original input, keep it undefined in the output
+        const currentVal = numbers[valueIndex]
+        if (currentVal === null) {
+            result[valueIndex] = InvalidCellTypes.NullButShouldBeNumber
+            continue
+        } else if (currentVal === undefined) {
+            result[valueIndex] = InvalidCellTypes.UndefinedButShouldBeNumber
+            continue
+        } else if (currentVal instanceof InvalidCell) {
+            result[valueIndex] = currentVal
+            continue
+        }
+
+        // Take away 1 for the current value (windowSize=1 means no smoothing & no expansion)
+        const expand = windowSize - 1
+
+        // With centered smoothing, expand uneven windows asymmetrically (ceil & floor) to ensure
+        // a correct number of window values get taken into account.
+        // Arbitrarily biased towards left (past).
+        const expandLeft = align === "center" ? Math.ceil(expand / 2) : expand
+        const expandRight = align === "center" ? Math.floor(expand / 2) : 0
+
+        const startIndex = Math.max(valueIndex - expandLeft, 0)
+        const endIndex = Math.min(valueIndex + expandRight, numbers.length - 1)
+
+        let count = 0
+        let sum = 0
+        for (
+            let windowIndex = startIndex;
+            windowIndex <= endIndex;
+            windowIndex++
+        ) {
+            const value = numbers[windowIndex]
+            if (
+                value !== undefined &&
+                value !== null &&
+                !(value instanceof InvalidCell)
+            ) {
+                sum += value!
+                count++
+            }
+        }
+
+        result[valueIndex] = sum / count
+    }
+
+    return result
+}
 
 // Assumptions: data is sorted by entity, then time
 // todo: move tests over from CE
